refactor: UserInfo 엔티티 패키지에서 dto.response로 이동 및 UserInfoResponseDTO로 이름 변경
