+++ src/main/java/com/wepong/pongdang/dto/response/LoginResponseDTO.java
@@ -1,6 +1,5 @@
 package com.wepong.pongdang.dto.response;
 
-import com.wepong.pongdang.entity.mapping.UserInfo;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class LoginResponseDTO {
     private String accessToken;
     private String message;
-    private UserInfo user;
+    private UserInfoResponseDTO user;
 }
